tms/requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
astor==0.7.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
DateTime==4.3
decorator==4.4.0
fire==0.1.3
fix-yahoo-finance==0.1.33
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.6
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
mock==3.0.4
multitasking==0.0.7
nbformat==4.4.0
numpy==1.16.3
pandas==0.24.2
pandas-datareader==0.7.0
plotly==3.8.1
protobuf==3.7.1
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
regex==2019.4.14
requests==2.21.0
retrying==1.3.3
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.15.2
wrapt==1.11.1
zope.interface==4.6.0
